delete CheCluster when CheInstallation is deleted
relying on the ownerRef does not work because the CheCluster
resource also has a finalizer which is handled by the che
operator, but when the CheInstallation resource is deleted,
the 'toolchain-workspaces' namespace is also deleted (by
ownership), which in turn removes the Che operator and if
this happens before the CheCluster resource was handled,
then the latter still exists and the parent namespace
is stuck in a 'Terminating' state.

Also, when the namespace is in 'Terminating' state (if
an admin deleted it manually), we delete the CheCluster
for the same reason.

Bonus: a test which was previously commented out is now
part of the E2e testsuite ;)

Fixes https://issues.redhat.com/browse/CRT-464

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>